timecoder: Use less needed valid bits for Traktor MK2
Since the Serato timecode needs only 20 bits to determine the new
position, the probability of bit errors and undesired jumps is quite
high. This is alleviated by artificially increasing the valid bits
needed for declaring the timecode valid to 44 by adding 24 bits.

Interestingly this is the exact number of bits the Traktor MK2s would
represent, if it had a 1kHz carrier signal as the Seratos do, since the
Traktor MK2s have a 2.5 kHz signal and need 110 bits:

100 / 2.5 = 44

This means that the recovery time of the Traktor MK2s after a needle
drop is equivalent to the one of Serato.