lut: Extend LUT with functions for the MK2 which can hold 110-bit LFSR states
Since the Traktor MK2 timecode is comprised of an LFSR, which generator
polynomial is of 110th order, a new type must be created to hold these
values.

This significantly increases the size of the LUT due to the type
being used in struct slot. Therefore the functions for the MK2 are
clearly separated to not make the current LUT gain in size by
holding unnecessarily padded zeros.