timecoder: Implement pitch detection
The Traktor MK2 signal is offset-modulated and therefore not directly
suited for pitch detection.

A suitable signal is the derivative, which more or less evenly oscillates
around zero. Since the derivative's amplitude is only a fraction of the
amplitude of the original signal, it needs proper scaling.

Scaling is achieved via computing the RMS value of the running sums of
the signal and derivative. The resulting quotient of RMS_signal /
RMS_derivative yields the needed scaling factor.