timecoder: Add internal mechanics for the Traktor MK2 timecode
At face value the Traktor MK2 LFSR seems to be of 110th order, i.e. it has a length of 110
bits. This results in a huge lookup table and requires an integer type that holds 110 bits
which has some computational and huge memory drawbacks. However, if we decimate the
sequence by a factor of five, meaning we only take every fifth bit into account, we can
observe that there are actually two 22-bit LFSRs hidden in this code, that repeat in the
following fashion as follows. The state of LFSR-A repeats three times, then the state of
LFSR-B repeats two times, then LFSR-A shifts forward and repeats three times again and so
forth.

The actual timecode slot can be computed as follows:

 Slot-A/B | A/B rep |       Function        | Slot
 ---------|---------|-----------------------|-----
   A = 0  |    0    | (a * 5) + rep         |  0    <-- LUT-A starts with Slot-B = 0
   A = 0  |    1    | (a * 5) + rep         |  1
   A = 0  |    2    | (a * 5) + rep         |  2
   B = 1  |    1    | ((b-1) * 5) + 3 + rep |  4    <-- LUT-B starts with Slot-B = 1
   B = 1  |    0    | ((b-1) * 5) + 3 + rep |  3
   A = 1  |    0    | (a * 5) + rep         |  5
   A = 1  |    1    | (a * 5) + rep         |  6
   A = 1  |    2    | (a * 5) + rep         |  7
   B = 2  |    0    | ((b-1) * 5) + 3 + rep |  8
   B = 2  |    1    | ((b-1) * 5) + 3 + rep |  9

 To achieve this a 110-bit window is needed of which every fifth bit is appended to a
 22-bit sequence. This 22-bit sequence is verified against LUT-A or LUT-B. In case of a
 match, the actual timecode slot can be computed mk2_compute_actual_slot() function.